add request type checking for delete_team_member view function

+++ app/views.py
@@ -387,13 +387,14 @@ def delete_team_member(request, member_id):
     # Use case : to delete single team member.
     # We are going to find the team member to delete by using it's (id)
 
-    # Fetching single team member
-    # get : will error in case the record does not exists.
-    # TODO : avoid this error
-    team_member = TeamMember.objects.get(id=member_id)
-    print(f"\n\nteam_member : {team_member}\n\n")
+    if request.method == "POST":
+        # Fetching single team member
+        # get : will error in case the record does not exists.
+        # TODO : avoid this error
+        team_member = TeamMember.objects.get(id=member_id)
+        print(f"\n\nteam_member : {team_member}\n\n")
 
-    # Deleting single team member
-    team_member.delete()
+        # Deleting single team member
+        team_member.delete()
 
     return redirect("get_team_members")
